Added unit tests for GeomagActivityEvaluator and GeomagLocationEvaluator

+++ app/src/main/kotlin/se/gustavkarlsson/skylight/android/services_impl/evaluation/GeomagLocationEvaluator.kt
@@ -1,9 +1,9 @@
 package se.gustavkarlsson.skylight.android.services_impl.evaluation
 
 import dagger.Reusable
-import se.gustavkarlsson.skylight.android.services.evaluation.Chance.Companion.UNKNOWN
 import se.gustavkarlsson.skylight.android.entities.GeomagLocation
 import se.gustavkarlsson.skylight.android.services.evaluation.Chance
+import se.gustavkarlsson.skylight.android.services.evaluation.Chance.Companion.UNKNOWN
 import se.gustavkarlsson.skylight.android.services.evaluation.ChanceEvaluator
 import javax.inject.Inject
 
@@ -14,7 +14,7 @@ constructor() : ChanceEvaluator<GeomagLocation> {
 
     override fun evaluate(value: GeomagLocation): Chance {
         val latitude = value.latitude ?: return UNKNOWN
-        val absoluteLatitude = Math.abs(latitude).toDouble()
+        val absoluteLatitude = Math.abs(latitude)
         var chance = 1.0 / 12.0 * absoluteLatitude - 55.0 / 12.0 // 55-67
         if (chance > 1.0) {
             chance = 2.0 - chance
